+++ CHANGELOG.md
@@ -7,7 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ----------------------
 ##### Fixed
 - Download links in file input and draw widgets not working in IE11.
-- The required asterisks (*) do not update in real-time.
+- The required asterisks (*) do not update in real-time if `validateContinously` is set to `false`.
 
 [4.55.1] - 2018-07-25
 ----------------------