+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-[Unreleased]
+[4.55.2] - 2018-08-02
 ----------------------
 ##### Fixed
 - Download links in file input and draw widgets not working in IE11.